refactor(map, menubutton): abstract ESC key press handler as custom hook
dry the code for closing a dialog with ESC key press
